OY-5342 Reduce transaction isolation on reads (#432)
In read transactions with more than one select query, use Read committed
level isolation, since full Serializable seems like overkill.

For the few transactional queries with only a single select, drop the
transactionality completely.

Refactor the `runBlockingTransactionally` to have a separate parameter
list for the optional parameters, to make it nicer to read when
supplying only some of them.

Also, add an index in hakukohteiden valintakokeet for hakukohde. This
speeds some some indexing queries, for koulutus with a big number of
hakukohteet.